Refactor simulate entrypoint and logging setup:
- Move log instantiation & reset into LogSettings.init_logs()
- Update simulate.py to call init_logs() and use master RNG
- Rename and clean up variables in simulate.py (game_seed_pairs, etc.)
- Extract game helper functions into monopoly/core/game_utils.py
- Tidy up analytics invocation to Analyzer().run_all()
- Adjust LogSettings paths and constants (KEEP_GAME_LOG, EVENTS_LOG_PATH)
